project folder structure changed to separate middlewares concerns
